Create NUMBER GUESSING GAME(after each guess the range changes)
After each guess the range changes and software shows new range. For example we have to guess a number between 0-10 and if the correct answer is 6 and we enter 4 so the compiler says that the new range is from 4-10. Now if we enter 8 so the compiler says that the new range is 4-8 and so on the range keeps on decreasing till we enter the correct answer. This makes it easy to guess the number. Also at last it displays the number of tries taken by the user.